updater/requirements.txt
requests==2.32.3
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.3.0
pyyaml==6.0.2
